# PERF: `Styler.tooltips` 