When tooltips were developed there were no hidden_rows or hidden_columns and no max_rows or max_columns.

This adds some benchmarks and a minimal change to improve performance. Scope to do a bit better here in future also.

```
       before           after         ratio
     [cd13e3a4]       [7766f055]
     <master>         <styler_tooltips_perf>
-        52.5±2ms       47.4±0.7ms     0.90  io.style.Render.time_apply_format_hide_render(24, 120)
-       196±0.6ms        173±0.9ms     0.88  io.style.Render.time_tooltips_render(36, 120)
-        78.0±2ms       67.0±0.4ms     0.86  io.style.Render.time_tooltips_render(12, 120)
-         140±2ms        120±0.3ms     0.85  io.style.Render.time_tooltips_render(24, 120)

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```